+++ src/http/request_context.rs
@@ -34,11 +34,65 @@ impl RequestContext {
   pub fn get_min_max_age(&self) -> Option<u64> {
     *self.min_max_age.lock().unwrap()
   }
+
+  pub fn update_max_age(&self, max_age: Option<std::time::Duration>) {
+    if let Some(ttl) = max_age {
+      let ttl_secs = ttl.as_secs();
+      let min_max_age_lock = self.get_min_max_age();
+      match min_max_age_lock {
+        Some(min_max_age) if ttl_secs < min_max_age => {
+          self.set_min_max_age(ttl_secs);
+        }
+        None => {
+          self.set_min_max_age(ttl_secs);
+        }
+        _ => {}
+      }
+    }
+  }
 }
 
 impl From<&ServerContext> for RequestContext {
   fn from(server_ctx: &ServerContext) -> Self {
     let http_client = server_ctx.http_client.clone();
     Self::new(http_client, server_ctx.server.clone())
+  }
+}
+
+#[cfg(test)]
+mod test {
+  use std::time::Duration;
+
+  use crate::http::RequestContext;
+
+  #[test]
+  fn test_update_max_age_none() {
+    let req_ctx = RequestContext::default();
+    req_ctx.set_min_max_age(120);
+    req_ctx.update_max_age(None);
+    assert_eq!(req_ctx.get_min_max_age(), Some(120));
+  }
+
+  #[test]
+  fn test_update_max_age_less_than_existing() {
+    let req_ctx = RequestContext::default();
+    req_ctx.set_min_max_age(120);
+    req_ctx.update_max_age(Some(Duration::new(60, 0)));
+    assert_eq!(req_ctx.get_min_max_age(), Some(60));
+  }
+
+  #[test]
+  fn test_update_max_age_greater_than_existing() {
+    let req_ctx = RequestContext::default();
+    req_ctx.set_min_max_age(60);
+    req_ctx.update_max_age(Some(Duration::new(120, 0)));
+    assert_eq!(req_ctx.get_min_max_age(), Some(60));
+  }
+
+  #[test]
+  fn test_update_max_age_no_existing_value() {
+    let req_ctx = RequestContext::default();
+    req_ctx.update_max_age(Some(Duration::new(120, 0)));
+    assert_eq!(req_ctx.get_min_max_age(), Some(120));
   }
 }
